Keep existing LifecycleInjector as is and move all the non root injector code into a new SimpleLifecycleInjector class
