+++ src/main/java/com/netflix/governator/guice/LifecycleInjector.java
@@ -60,12 +60,10 @@ public class LifecycleInjector
     private final List<Module> modules;
     private final Collection<Class<?>> ignoreClasses;
     private final boolean ignoreAllClasses;
-    private LifecycleManager lifecycleManager;
-    private Injector injector;
+    private final LifecycleManager lifecycleManager;
+    private final Injector injector;
     private final List<Module> discoveredModules = Lists.newArrayList();
-    private final List<BootstrapModule> bootstrapModules;
-    private Stage stage;
-    
+
     /**
      * Create a new LifecycleInjector builder
      *
@@ -194,42 +192,16 @@ public class LifecycleInjector
         if ( !ignoreAllClasses )
         {
             Collection<Class<?>>    localIgnoreClasses = Sets.newHashSet(ignoreClasses);
-            localModules.add(new InternalAutoBindModule(
-                injector == null 
-                ? Guice.createInjector(stage, new InternalBootstrapModule(this.scanner, bootstrapModules)) 
-                : injector, scanner, localIgnoreClasses));
+            localModules.add(new InternalAutoBindModule(injector, scanner, localIgnoreClasses));
         }
-        
-        if (injector == null) {
-            AtomicReference<LifecycleManager> lifecycleManagerRef = new AtomicReference<LifecycleManager>();
-            
-            List<Module> fullModuleList = Lists.newArrayList(
-                new InternalBootstrapModule(this.scanner, bootstrapModules),
-                new InternalLifecycleModule(lifecycleManagerRef)
-                );
-            
-            fullModuleList.addAll(localModules);
 
-            injector = Guice.createInjector
-            (
-                stage,
-                fullModuleList
-            );
-            
-            lifecycleManager = lifecycleManagerRef.get();
-            return injector;
-        }
-        else {
-            return createChildInjector(localModules);
-        }
+        return createChildInjector(localModules);
     }
 
-    LifecycleInjector(List<Module> modules, Collection<Class<?>> ignoreClasses, boolean ignoreAllClasses, List<BootstrapModule> bootstrapModules, ClasspathScanner scanner, Collection<String> basePackages, Stage stage, Class<?> rootModule, boolean discardBootstrapInjector)
+    LifecycleInjector(List<Module> modules, Collection<Class<?>> ignoreClasses, boolean ignoreAllClasses, List<BootstrapModule> bootstrapModules, ClasspathScanner scanner, Collection<String> basePackages, Stage stage, Class<?> rootModule)
     {
         stage = Preconditions.checkNotNull(stage, "stage cannot be null");
 
-        this.stage = stage;
-        this.bootstrapModules = bootstrapModules;
         this.ignoreAllClasses = ignoreAllClasses;
         this.ignoreClasses = ImmutableList.copyOf(ignoreClasses);
         this.modules = ImmutableList.copyOf(modules);
@@ -237,26 +209,17 @@ public class LifecycleInjector
         
         InternalModuleDependencyModule moduleDepdencyModule = new InternalModuleDependencyModule();
         AtomicReference<LifecycleManager> lifecycleManagerRef = new AtomicReference<LifecycleManager>();
-
-        if (discardBootstrapInjector) {
-            if (rootModule != null) {
-                Injector tempInjector = Guice.createInjector(moduleDepdencyModule);
-                tempInjector.getInstance(rootModule);
-                this.discoveredModules.addAll(moduleDepdencyModule.getModules());
-            }
-        }
-        else {
-            injector = Guice.createInjector
-            (
-                stage,
-                new InternalLifecycleModule(lifecycleManagerRef),
-                new InternalBootstrapModule(this.scanner, bootstrapModules),
-                moduleDepdencyModule
-            );
-            lifecycleManager = lifecycleManagerRef.get();
-            if (rootModule != null)
-                injector.getInstance(rootModule);
-            this.discoveredModules.addAll(moduleDepdencyModule.getModules());
-        }
+        injector = Guice.createInjector
+        (
+            stage,
+            new InternalBootstrapModule(this.scanner, bootstrapModules),
+            new InternalLifecycleModule(lifecycleManagerRef),
+            moduleDepdencyModule
+        );
+        if (rootModule != null)
+            injector.getInstance(rootModule);
+        this.discoveredModules.addAll(moduleDepdencyModule.getModules());
+        lifecycleManager = injector.getInstance(LifecycleManager.class);
+        lifecycleManagerRef.set(lifecycleManager);
     }
 }
